pyproject.toml
[tool.poetry]
name = "protouch"
version = "0.1.0"
description = ""
authors = ["Komilov Shakhzod <proap7875@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
python-dotenv = "^1.1.0"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
django-jazzmin = "^3.0.1"
pillow = "^11.2.1"
django-cors-headers = "^4.7.0"
django-mptt = "^0.17.0"
djangorestframework-simplejwt = "^5.5.0"
requests = "^2.32.3"
django-filter = "^25.1"
social-auth-app-django = "^5.4.3"
django-modeltranslation = "^0.19.14"
sentry-sdk = {extras = ["django"], version = "^2.29.1"}
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
